make it work on Windows

+++ include/binary_data_source.h
@@ -38,7 +38,7 @@ public:
 
     struct stat sb;
     stat(path.c_str(), &sb);
-    n = std::min(sb.st_size / (sizeof(float) * d_), n_);
+    n = std::min((size_t)sb.st_size / (sizeof(float) * d_), n_);
     d = d_;
 
     data = new ElementType[n * d];
